Move all config into init.el, remove helm
It's just config after all, it should not grow too big. It's
conveniece to just fit into one file, it's not worth to separate into
multiple files and spent too much time maintaining.

Helm's TAB behavior was not Emacs-style, I want consistent
mini-buffer, so just use built-in ido-mode. Another reason to remove
helm is that it was slow when use helm-etags-select on large project
such as linux kernel.